+++ packages/terra-core-docs/CHANGELOG.md
@@ -20,6 +20,7 @@
   * Added `level` prop to `terra-action-header` tests and examples.
   * Added upgrade guide for `terra-action-header`.
   * Added Accessibility guide for `terra-button`.
+  * Changes to Accessibility guide for `terra-button`.
 
 * Fixed
   * Updated broken links in terra-form-select and action-footer.